move test to api package, break down cache key method, change to Truth testing
