+++ apollo-api/build.gradle
@@ -6,6 +6,9 @@ sourceCompatibility = JavaVersion.VERSION_1_7
 dependencies {
   compile dep.jsr305
   compile dep.jsr250
+
+  testCompile dep.junit
+  testCompile dep.truth
 }
 
 apply from: rootProject.file('gradle/gradle-mvn-push.gradle')
